refactor(PlayerVisualizer): split controllers, extracted button and slider, separated hooks into individual files
