+++ src/modules/player/components/PlayerVisualizer/controls/AudioControls.tsx
@@ -3,10 +3,7 @@ type AudioControlsProps = {
   onPause: () => void;
   onStop: () => void;
   isPlaying: boolean;
-  volume: number;
-  onVolumeChange: (volume: number) => void;
-  isMuted: boolean;
-  onToggleMute: () => void;
+  extraControls?: React.ReactNode;
 };
 
 export const AudioControls: React.FC<AudioControlsProps> = ({
@@ -14,10 +11,7 @@ export const AudioControls: React.FC<AudioControlsProps> = ({
   onPause,
   onStop,
   isPlaying,
-  volume,
-  onVolumeChange,
-  isMuted,
-  onToggleMute
+  extraControls
 }) => {
   return (
     <div className="flex flex-col items-center gap-6 mt-6">
@@ -45,25 +39,7 @@ export const AudioControls: React.FC<AudioControlsProps> = ({
         </button>
       </div>
 
-      <div className="flex items-center gap-4 mt-4">
-        <span className="text-gray-600">Volume</span>
-        <input
-          type="range"
-          min="0"
-          max="1"
-          step="0.01"
-          value={volume}
-          onChange={(e) => onVolumeChange(Number(e.target.value))}
-          className="w-64 h-2 bg-gray-300 rounded-lg accent-blue-500 transition-all"
-        />
-      </div>
-
-      <button
-        onClick={onToggleMute}
-        className="mt-4 px-6 py-3 bg-gray-500 text-white rounded-lg shadow-lg transform transition-all duration-200 hover:bg-gray-400"
-      >
-        {isMuted ? "Unmute" : "Mute"}
-      </button>
+      {extraControls}
     </div>
   );
 };
