+++ src/modules/player/components/PlayerVisualizer/hooks/useAudioPlayer.ts
@@ -1,0 +1,49 @@
+import { useState, useRef } from "react";
+
+export const useAudioPlayer = () => {
+  const [currentTime, setCurrentTime] = useState<number>(0);
+  const [isPlaying, setIsPlaying] = useState<boolean>(false);
+
+  const audioRef = useRef<HTMLAudioElement | null>(null);
+
+  const playAudio = () => {
+    if (audioRef.current) {
+      audioRef.current.play();
+      setIsPlaying(true);
+    }
+  };
+
+  const pauseAudio = () => {
+    if (audioRef.current) {
+      audioRef.current.pause();
+      setIsPlaying(false);
+    }
+  };
+
+  const stopAudio = () => {
+    if (audioRef.current) {
+      audioRef.current.pause();
+      audioRef.current.currentTime = 0;
+      setIsPlaying(false);
+      setCurrentTime(0);
+    }
+  };
+
+  const handleSeek = (time: number) => {
+    if (audioRef.current) {
+      audioRef.current.currentTime = time;
+      setCurrentTime(time);
+    }
+  };
+
+  return {
+    currentTime,
+    setCurrentTime,
+    isPlaying,
+    audioRef,
+    playAudio,
+    pauseAudio,
+    stopAudio,
+    handleSeek
+  };
+};
